# api.yaml
openapi: 3.0.4
info:
  title: E-commerce App API
  description: |-
    This is a sample E-commerce App Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the E-commerce App, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [The E-commerce App repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the E-commerce App](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: https://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
externalDocs:
  description: Find out more about Swagger
  url: https://swagger.io
servers:
  - url: http://localhost:3000/api
    description: Local server
tags:
  - name: customer
    description: Everything about your Customers
  - name: admin
    description: Operations about admin
  - name: product
    description: Access to Product
  - name: order
    description: Operations about Orders
  - name: auth 
    description: Operations about user authentication
paths:
  /products:
    get:
      tags:
        - product
      summary: Get all products
      description: Retrieve a list of all products with their details, including category and images.
      operationId: getProducts
      responses:
        '200':
          description: A list of products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSummary'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post: 
      tags:
        - product
      summary: Create a new product
      description: Add a new product to the catalog.
      operationId: createProduct
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSummary'
    
  /product/{id}:
    get:
      tags:
        - product
      summary: Get product by ID
      description: Retrieve detailed information about a specific product by its ID, including category and images.
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product to retrieve.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSummary'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put: 
      tags:
        - product
      summary: Update an existing product
      description: Update the details of an existing product.
      operationId: updateProduct
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSummary'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - product
      summary: Delete a product
      description: Remove a product from the catalog by its ID.
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the product to delete.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /product/search:
    get:
      tags:
        - product
      summary: Search products
      description: Search for products by name or description.
      operationId: searchProducts
      parameters:
        - name: query
          in: query
          required: true
          description: The search term to filter products by name or description.
          schema:
            type: string
      responses:
        '200':
          description: A list of products matching the search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSummary'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /product/categories:
    get:
      tags:
        - product
      summary: Get all categories
      description: Retrieve a list of all product categories.
      operationId: getCategories
      responses:
        '200':
          description: A list of categories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/customers:
    get:
      tags:
        - customer
      summary: Get all customers
      description: Retrieve a list of all customers.
      operationId: getCustomers
      responses:
        '200':
          description: A list of customers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerDetails'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/customer/{id}:
    get:
      tags:
        - customer
      summary: Get customer by ID
      description: Retrieve detailed information about a specific customer by their ID.
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the customer to retrieve.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetails'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders: 
    get:
      tags:
        - order
      summary: Get all orders
      description: Retrieve a list of all orders with their details, including customer and order items.
      operationId: getOrders
      responses:
        '200':
          description: A list of orders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    CustomerDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        full_name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        google_id:
          type: string
          example: 1234567890
        facebook_id:
          type: string
          description: Facebook ID
          example: 0987654321
        password:
          type: string
          format: password
          example: strongpassword123
        phone: 
          type: string
          example: '0123456789'
        address: 
          type: string
          example: '123 Main St, City, Country'
        created_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
      required:
        - full_name
        - email
        - password
    CustomerInput:
      type: object
      properties:
        full_name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        google_id:
          type: string
          example: 1234567890
        facebook_id:
          type: string
          description: Facebook ID
          example: 0987654321
        password:
          type: string
          format: password
          example: strongpassword123
        phone:
          type: string
          example: '0123456789'
        address:
          type: string
          example: '123 Main St, City, Country'
      required:
        - full_name
        - email
        - password
        - phone
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Sofa
        description:
          type: string
          example: A comfortable sofa
        created_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
    Product_Image:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        product_id:
          type: integer
          format: int64
          example: 10
        url:
          type: string
          format: uri
          example: https://example.com/images/product1.jpg
        created_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        Product:
          $ref: '#/components/schemas/ProductSummary'
    ProductDetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Sofa
        description:
          type: string
          example: A comfortable sofa
        price:
          type: number
          format: float
          example: 299.99
        sale_price:
          type: number
          format: float
          example: 249.99
        category_id:
          type: number
          format: int64
          example: 1
        stock_quantity:
          type: number
          format: int64
          example: 36
        material:
          type: string
          description: Material of the product
          example: Leather
        color:
          type: string
          description: Color of the product
          example: Red
        dimensions:
          type: string
          description: Dimensions of the product
          example: 200x80x100 cm
        status:
          type: string
          description: Status of the product
          enum:
            - active
            - inactive
          example: active
        created_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        category:
          $ref: '#/components/schemas/Category'
        images:
          type: array
          items:
            $ref: '#/components/schemas/Product_Image'
        order_details:
          type: array
          items:
            $ref: '#/components/schemas/Order_Details'
    ProductInput:
      type: object
      properties:
        name:
          type: string
          example: Sofa
        description:
          type: string
          example: A comfortable sofa
        price:
          type: number
          format: float
          example: 299.99
        sale_price:
          type: number
          format: float
          example: 249.99
        category_id:
          type: number
          format: int64
          example: 1
        stock_quantity:
          type: number
          format: int64
          example: 36
        material:
          type: string
          description: Material of the product
          example: Leather
        color:
          type: string
          description: Color of the product
          example: Red
        dimensions:
          type: string
          description: Dimensions of the product
          example: 200x80x100 cm
        images:
          type: array
          items:
            $ref: '#/components/schemas/Product_Image'
      required:
        - name
        - price
        - category_id
        - stock_quantity
        - images
    ProductSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Sofa
        price:
          type: number
          format: float
          example: 299.99
        sale_price:
          type: number
          format: float
          example: 249.99
        category_id:
          type: number
          format: int64
          example: 1
        stock_quantity:
          type: number
          format: int64
          example: 36
        status:
          type: string
          description: Status of the product
          enum:
            - active
            - inactive
          example: active
        category:
          $ref: '#/components/schemas/Category'
        images:
          type: array
          items:
            $ref: '#/components/schemas/Product_Image'
    Order_Details:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        order_id:
          type: integer
          format: int64
          example: 10
        product_id:
          type: integer
          format: int64
          example: 20
        quantity:
          type: integer
          format: int32
          example: 2
        unit_price:
          type: number
          format: float
          example: 299.99
        total_price:
          type: number
          format: float
          example: 599.98
        order: 
          $ref: '#/components/schemas/Order'
        product:
          $ref: '#/components/schemas/ProductSummary'
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        order_number:
          type: string
          example: ORD-1001
        customer_id:
          type: integer
          format: int64
          example: 5
        order_date:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        total_amount:
          type: number
          format: float
          example: 599.98
        shipping_fee:
          type: number
          format: float
          example: 15.00
        status:
          type: string
          description: Order status
          enum:
            - pending
            - confirmed
            - shipping
            - completed
            - cancelled
          example: pending
        payment_method:
          type: string
          description: Payment method
          enum:
            - cash_on_delivery
            - bank_transfer
            - credit_card
            - e_wallet
            - installment
          example: cash_on_delivery
        payment_status:
          type: string
          description: Payment status
          enum:
            - pending
            - paid
            - failed
            - refunded
          example: pending
        paid_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        shipping_method:
          type: string
          description: Shipping method
          enum:
            - standard_delivery
            - express_delivery
            - same_day_delivery
            - pickup_at_store
            - installation_service
          example: standard_delivery
        shipping_address:
          type: string
          example: '123 Main St, City, Country'
        phone:
          type: string
          example: '0123456789'
        expected_delivery_date:
          type: string
          format: date-time
          example: '2023-10-05T12:00:00Z'
        notes:
          type: string
          example: 'Please deliver between 9 AM and 5 PM.'
        created_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-10-01T12:00:00Z'
        customer:
          $ref: '#/components/schemas/CustomerDetails'
        order_details:
          type: array
          items:
            $ref: '#/components/schemas/Order_Details'
        
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: tag
    Pet:
      required:
        - name
        - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: doggie
        category:
          $ref: '#/components/schemas/Category'
        photoUrls:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            xml:
              name: photoUrl
        tags:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        status:
          type: string
          description: pet status in the store
          enum:
            - available
            - pending
            - sold
      xml:
        name: pet
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            "write:pets": modify pets in your account
            "read:pets": read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header